Update full (perfect) scoreboard
Scoreboard is fully implemented, with  the arbitrary length of its board(max = 5), depends on userInput. It deeply rooted in user's Phone memory So it can be accessed at any time even if you closed the app. GameOverpop is updated. Mainfest file is added permission on read and write into the external memory.